MAINT: move util password to database (#225)
