+++ doc/api.rst
@@ -68,6 +68,7 @@ Submission-related tables
    model.SubmissionFile
    model.SubmissionFileType
    model.SubmissionOnCVFold
+   model.SubmissionScoreOnCVFold
    model.SubmissionSimilarity
 
 Relationship tables
@@ -314,6 +315,8 @@ Frontend-related database tools
    :toctree: generated/
    :template: function.rst
 
+   utils.check_password
+   utils.hash_password
    utils.setup_db
    utils.session_scope
 
@@ -564,5 +567,4 @@ RAMP shared utilities
    :template: function.rst
 
    import_module_from_source
-   utils.check_password
-   utils.hash_password
+
