+++ ramp-database/ramp_database/tools/tests/test_user.py
@@ -4,10 +4,10 @@ import pandas as pd
 import pytest
 
 from ramp_utils import read_config
-from ramp_utils.utils import check_password
 from ramp_utils.testing import database_config_template
 from ramp_utils.testing import ramp_config_template
 
+from ramp_database.utils import check_password
 from ramp_database.utils import setup_db
 from ramp_database.utils import session_scope
