+++ ramp-database/ramp_database/utils.py
@@ -5,6 +5,8 @@ the RAMP database.
 
 from contextlib import contextmanager
 
+import bcrypt
+
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.engine.url import URL
@@ -67,3 +69,43 @@ def session_scope(config):
             raise
         finally:
             session.close()
+
+
+def _encode_string(text):
+    return bytes(text, 'utf-8') if isinstance(text, str) else text
+
+
+def hash_password(password):
+    """Hash a password.
+
+    Parameters
+    ----------
+    password : str or bytes
+        Human readable password.
+
+    Returns
+    -------
+    hashed_password : bytes
+        The hashed password.
+    """
+    return bcrypt.hashpw(_encode_string(password), bcrypt.gensalt())
+
+
+def check_password(password, hashed_password):
+    """Check if a password is the same than the hashed password.
+
+    Parameters
+    ----------
+    password : str or bytes
+        Human readable password.
+    hashed_password : str or bytes
+        The hashed password.
+
+    Returns
+    -------
+    is_same_password : bool
+        Return True if the two passwords are identical.
+    """
+    return bcrypt.checkpw(
+        _encode_string(password), _encode_string(hashed_password)
+    )
