+++ ramp-utils/ramp_utils/utils.py
@@ -1,7 +1,5 @@
 import importlib
 
-import bcrypt
-
 
 def import_module_from_source(source, name):
     """Load a module from a Python source file.
@@ -21,43 +19,3 @@ def import_module_from_source(source, name):
     module = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(module)
     return module
-
-
-def _encode_string(text):
-    return bytes(text, 'utf-8') if isinstance(text, str) else text
-
-
-def hash_password(password):
-    """Hash a password.
-
-    Parameters
-    ----------
-    password : str or bytes
-        Human readable password.
-
-    Returns
-    -------
-    hashed_password : bytes
-        The hashed password.
-    """
-    return bcrypt.hashpw(_encode_string(password), bcrypt.gensalt())
-
-
-def check_password(password, hashed_password):
-    """Check if a password is the same than the hashed password.
-
-    Parameters
-    ----------
-    password : str or bytes
-        Human readable password.
-    hashed_password : str or bytes
-        The hashed password.
-
-    Returns
-    -------
-    is_same_password : bool
-        Return True if the two passwords are identical.
-    """
-    return bcrypt.checkpw(
-        _encode_string(password), _encode_string(hashed_password)
-    )
